feat: 기이수과목 커스텀 수정 (#53)

+++ core/views.py
@@ -21,16 +21,17 @@ def mypage(request):
 
 @login_required
 def custom(request):
-    courses = Course.objects.filter(user=request.user).order_by('-pk')
-    context = {'courses':courses}
+    courses = Course.objects.filter(user=request.user, custom=False).order_by('-pk')
+    customs = Course.objects.filter(user=request.user, custom=True).order_by('-pk')
+    context = {'courses': courses, 'customs': customs}
     return render(request, 'core/custom.html', context)
 
 @login_required
 def course_delete(request, course_id):
     course = get_object_or_404(Course, pk=course_id)
-    if request.user != course.user:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('core:custom', course_id=course.subject.number)
+    if request.user != course.user or not course.custom:
+        messages.error(request, '삭제권한이 없습니다.')
+        return redirect('core:custom')
     course.delete()
     return redirect('core:custom')
